config has default and a few tests fixed and others made much more robust to error

+++ tests/modules/job_control/test_models.py
@@ -10,32 +10,36 @@ def test_job_control(flask_app, researcher_1, test_root, db):
     from app.modules.job_control.models import JobControl
     from app.modules.asset_groups.tasks import sage_detection
 
-    asset_group = AssetGroup(owner_guid=researcher_1.guid)
-    ags = AssetGroupSighting(
-        asset_group_guid=asset_group.guid, config={'assetReferences': []}
-    )
+    asset_group = None
+    try:
+        asset_group = AssetGroup(owner_guid=researcher_1.guid)
+        ags = AssetGroupSighting(
+            asset_group_guid=asset_group.guid, config={'assetReferences': []}
+        )
 
-    db.session.add(asset_group)
-    db.session.add(ags)
+        db.session.add(asset_group)
+        db.session.add(ags)
 
-    with mock.patch(
-        'app.modules.asset_groups.models.current_app', return_value=flask_app
-    ):
-        utc_now = datetime.datetime(2021, 6, 29, 8, 22, 35)
-        with mock.patch('datetime.datetime') as mock_datetime:
-            mock_datetime.utcnow.return_value = utc_now
-            sage_detection(str(ags.guid), 'african_terrestrial')
-        job_id = list(ags.jobs.keys())[0]
+        with mock.patch('app.modules.asset_groups.models.current_app') as mock_app:
+            mock_app.return_value = flask_app
+            mock_app.config.get.return_value = 'zebra'
+            utc_now = datetime.datetime(2021, 6, 29, 8, 22, 35)
+            with mock.patch('datetime.datetime') as mock_datetime:
+                mock_datetime.utcnow.return_value = utc_now
+                sage_detection(str(ags.guid), 'african_terrestrial')
 
-        # TODO asserts
-        JobControl.check_jobs()
+            job_id = list(ags.jobs.keys())[0]
 
-        with mock.patch('app.modules.asset_groups.models.log') as log:
-            JobControl.print_jobs()
-            assert log.warning.call_args_list, [
-                mock.call(
-                    f'AssetGroupSighting:{ags.guid} Job:{job_id} Model:Animal UTC Start:{utc_now}'
-                ),
-            ]
+            # TODO asserts
+            JobControl.check_jobs()
 
-    asset_group.delete()
+            with mock.patch('app.modules.asset_groups.models.log') as log:
+                JobControl.print_jobs()
+                assert log.warning.call_args_list, [
+                    mock.call(
+                        f'AssetGroupSighting:{ags.guid} Job:{job_id} Model:Animal UTC Start:{utc_now}'
+                    ),
+                ]
+    finally:
+        if asset_group:
+            asset_group.delete()
